requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
colorama==0.4.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
